Increase coverage for collapsible flashbar with animations
